proc: Fix parsing Kernel Module lines where refcount is "-"

+++ interpreter/loaderinfo.go
@@ -55,7 +55,7 @@ func (i *LoaderInfo) GetSymbolAsRanges(symbol libpf.SymbolName) ([]util.Range, e
 	start := uint64(sym.Address)
 	return []util.Range{{
 		Start: start,
-		End:   start + uint64(sym.Size)},
+		End:   start + sym.Size},
 	}, nil
 }
